requirements.txt
streamlit
openai
numpy
hugchat
python-dotenv
